Implement Tasks policies

+++ app/Http/Controllers/Tasks/DeleteTask.php
@@ -4,8 +4,10 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers\Tasks;
 
+use App\Exceptions\TaskNotFoundException;
 use App\Http\Controllers\Controller;
 use App\Services\TasksService;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 
 class DeleteTask extends Controller
@@ -14,8 +16,15 @@ class DeleteTask extends Controller
     {
     }
 
+    /**
+     * @throws AuthorizationException
+     * @throws TaskNotFoundException
+     */
     public function __invoke(int $id): JsonResponse
     {
+        $taskVo = $this->tasksService->get($id);
+        $this->authorize('delete', [$taskVo]);
+
         $this->tasksService->delete($id);
 
         return $this->emptyResponse();
